% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_analyses}
\alias{add_analyses}
\title{Add analyses}
\usage{
add_analyses(x, x2)
}
\arguments{
\item{x}{An mb_analysis or mb_analyses object.}

\item{x2}{n mb_analysis or mb_analyses object.}
}
\value{
An object of class mb_analyses.
}
\description{
Add analyses
}
