% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_models}
\alias{add_models}
\title{Add model(s)}
\usage{
add_models(x, x2)
}
\arguments{
\item{x}{An mb_model or mb_models object.}

\item{x2}{n mb_model or mb_models object.}
}
\value{
An object of class mb_models.
}
\description{
Add model(s)
}
