% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{AIC.mb_analysis}
\alias{AIC.mb_analysis}
\title{Akaike's Information Criteron}
\usage{
\method{AIC}{mb_analysis}(object, n = NULL, ...)
}
\arguments{
\item{object}{The object to calculate the AIC for.}

\item{n}{A count of the sample size.}

\item{...}{Not used.}
}
\value{
The Akaike's Information Criteron as a number.
}
\description{
Calculates marginal AICc for an analysis.
}
\details{
By default the sample size is assumed to be the number of rows
in the analysis data set. To override manually set n to be an integer ie n = 10L.
To calculate AIC (as opposed to AICc) set n = Inf.
}

