% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{AICw}
\alias{AICw}
\title{Akaike's Information Criterion Weight}
\usage{
AICw(x, ...)
}
\arguments{
\item{x}{The object to calculate it for.}

\item{...}{Not used.}
}
\value{
A vector of the w_i values for each model.
}
\description{
Calculates w_i for a list of objects of class mb_analysis that return AIC.
}

