% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-data.R
\name{analyse_data}
\alias{analyse_data}
\title{Analyse Data}
\usage{
analyse_data(data, model, quick = getOption("mb.quick", FALSE),
  quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep", TRUE),
  ...)
}
\arguments{
\item{data}{The data frame to analyse.}

\item{model}{The model to use for the analysis.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Analyse data using an object inheriting from class mb_model.
}

