% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-priors.R
\name{analyse_priors}
\alias{analyse_priors}
\title{Analyse Priors}
\usage{
analyse_priors(analysis, multipliers = c(0.5, 2),
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep",
  TRUE), ...)
}
\arguments{
\item{analysis}{The analysis object with the priors to analyse.}

\item{multipliers}{A numeric vector of values to multiply the standard deviation of simple normal prior distributions by.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Analyse the priors to determine if they are influencing the posteriors.
}
