% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{analyse_residuals}
\alias{analyse_residuals}
\title{Analyse Residuals}
\usage{
analyse_residuals(x)
}
\arguments{
\item{x}{The mb_analysis object to analyse the residuals for.}
}
\description{
Analyse Residuals
}
