% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{analyses}
\alias{analyses}
\title{MB Analyses}
\usage{
analyses(...)
}
\arguments{
\item{...}{Named objects.}
}
\value{
An object inheriting from class mb_analyses.
}
\description{
Creates an object inherting from class mb_analyses.
}
