% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-priors.R
\name{analysis_priors}
\alias{analysis_priors}
\title{Analysis Priors}
\usage{
analysis_priors(object, ...)
}
\arguments{
\item{object}{The analysis object with the priors to return.}

\item{...}{Additional arguments.}
}
\description{
If present returns a list of the analysis priors. Otherwise returns NULL.
}
