% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayize.R
\name{arrayize}
\alias{arrayize}
\title{Arrayize}
\usage{
arrayize(vector, factors)
}
\arguments{
\item{vector}{The vector.}

\item{factors}{A named list of one or more factors.}
}
\value{
An array.
}
\description{
Converts a vector into a matrix or array based on two or more factors.
}
\details{
The function can be used in modify_data functions to convert vectors to matrices or
arrays.
}

