% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.analyses}
\alias{as.analyses}
\title{Coerce to an mb_analyses object}
\usage{
as.analyses(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerce to an mb_analyses object
}
