% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards.R
\name{backwards}
\alias{backwards}
\title{Backwards}
\usage{
backwards(model, data, drops = list(), conf_level = 0.95,
  beep = getOption("mb.beep", TRUE), ...)
}
\arguments{
\item{model}{An object.}

\item{data}{Data.}

\item{drops}{A list of character vectors specifying the scalar parameters to consider.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Unused.}
}
\value{
A list of the analyses.
}
\description{
Perform backwards step-wise regression on a model.
Returns a list of the analysis at each step starting with the full model.
}
\details{
drop is a list of character vectors specifying the scalar parameters to possibly drop.
If a list element consists of two or more strings then the earlier strings
are only available to drop after the later strings have been eliminated.
This allows polynomial dependencies to be respected.
}
