% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{base_model}
\alias{base_model}
\title{Base Model}
\usage{
base_model(model, drops = list())
}
\arguments{
\item{model}{The full model.}

\item{drops}{A list of character vectors specifying possible drops.}
}
\value{
The base model (with all drops).
}
\description{
Base Model
}
