% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mb_code}
\alias{check_mb_code}
\title{Check MB Code}
\usage{
check_mb_code(object, object_name = substitute(object))
}
\arguments{
\item{object}{The object to check.}

\item{object_name}{A string of the object name.}
}
\value{
The object or throws an informative error.
}
\description{
Check MB Code
}
