% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-model-parameters.R
\name{check_model_parameters}
\alias{check_model_parameters}
\title{Check Model Parameters}
\usage{
check_model_parameters(x, fixed, random, derived, drops)
}
\arguments{
\item{x}{The model code to check.}

\item{fixed}{A string of a regular expression specifying the fixed parameters to monitor.}

\item{random}{NULL or a character vector of the random effects.}

\item{derived}{NULL or a character vector of the derived parameters.}

\item{drops}{NULL or a character vector of the parameters to drop.}
}
\value{
If passes all checks, an invisible copy of x. Otherwise throws an informative error.
}
\description{
Check Model Parameters
}
