% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_uniquely_named_list}
\alias{check_uniquely_named_list}
\title{Check Uniquely Named List}
\usage{
check_uniquely_named_list(x, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the objects name.}
}
\value{
The original object or throws an informative error.
}
\description{
Check Uniquely Named List
}

