% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code}
\alias{code}
\title{Code}
\usage{
code(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Additional arguments.}
}
\value{
An object inheriting from class mb_code.
}
\description{
Gets the MB code for an object.
}
