% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.list}
\alias{coef.list}
\title{Coef List}
\usage{
\method{coef}{list}(object, param_type = "fixed", include_constant = TRUE,
  conf_level = 0.95, ...)
}
\arguments{
\item{object}{A list of mb_analysis objects.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coeffcient terms with the model averaged estimate, the
weight and the proportion of models including the term.
}
\description{
Coefficients for fixed parameters from
MB analyses averaged by IC weights.
}
