% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.lmb_analysis}
\alias{coef.lmb_analysis}
\title{Coef LMB Analysis}
\usage{
\method{coef}{lmb_analysis}(object, param_type = "fixed",
  include_constant = TRUE, conf_level = 0.95, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
Coefficients for a LMB analysis.
}
\details{
The (95\%) \code{lower} and \code{upper} confidence intervals are
the \code{estimate} +/- 1.96 * \code{std.error}.
}
