% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mb_analyses}
\alias{coef.mb_analyses}
\title{Coef TMB Analyses}
\usage{
\method{coef}{mb_analyses}(
  object,
  param_type = "fixed",
  include_constant = TRUE,
  conf_level = getOption("mb.conf_level", 0.95),
  estimate = getOption("mb.estimate", median),
  ...
)
}
\arguments{
\item{object}{The mb_analyses object.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{estimate}{The function to use to calculate the estimate for Bayesian models.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coeffcient terms with the model averaged estimate, the
Akaike's weight and the proportion of models including the term.
}
\description{
Coefficients for fixed parameters from an ML based
MB analyses averaged by IC weights.
}
