% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mb_analysis}
\alias{coef.mb_analysis}
\title{Coef JAGS Analysis}
\usage{
\method{coef}{mb_analysis}(
  object,
  param_type = "fixed",
  include_constant = TRUE,
  conf_level = getOption("mb.conf_level", 0.95),
  estimate = getOption("mb.estimate", median),
  simplify = TRUE,
  directional_information = FALSE,
  ...
)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{estimate}{The function to use to calculating the estimate for Bayesian models.}

\item{simplify}{Must be \code{TRUE}.
\code{simplify = FALSE} is defunct as of embr 1.1.0.}

\item{directional_information}{A flag specifying whether the svalue column
for a Bayesian analysis should be calculated using
\code{\link[extras:directional_information]{extras::directional_information()}} instead of \code{\link[extras:svalue]{extras::svalue()}}.
The default value will change from \code{FALSE} to \code{TRUE} in a future release;
set the argument explicitly to avoid the deprecation warning.
Ignored for frequentist analyses where the svalue is always \code{-log2(pvalue)}.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms with the columns indicating
the \code{term}, \code{estimate}, \code{lower} and \code{upper} confidence or credible intervals
and \code{svalue}.
}
\description{
Coefficients for a JAGS analysis.
}
