% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mb_meta_analyses}
\alias{coef.mb_meta_analyses}
\title{Coef TMB Meta Analyses}
\usage{
\method{coef}{mb_meta_analyses}(object, param_type = "fixed",
  include_constant = TRUE, conf_level = getOption("mb.conf_level", 0.95),
  estimate = getOption("mb.estimate", median),
  profile = getOption("mb.profile", FALSE), ...)
}
\arguments{
\item{object}{The mb_meta_analyses object.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{estimate}{The function to use to calculate the estimate.}

\item{profile}{A flag indicating whethe to use likelihood profiling as opposed to the Wald approximation when calculating confidence intervals for the frequentist models.}

\item{...}{Not used.}
}
\value{
A tidy tibble.
}
\description{
Coef TMB Meta Analyses
}
