% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mb_meta_analysis}
\alias{coef.mb_meta_analysis}
\title{Coef TMB Meta Analysis}
\usage{
\method{coef}{mb_meta_analysis}(object, param_type = "fixed",
  include_constant = TRUE, conf_level = getOption("mb.conf_level", 0.95),
  ...)
}
\arguments{
\item{object}{The mb_meta_analysis object.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Not used.}
}
\value{
A tidy tibble.
}
\description{
Coef TMB Meta Analysis
}
