% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-profile.R
\name{coef_profile.mb_meta_analysis}
\alias{coef_profile.mb_meta_analysis}
\title{Coef TMB Meta Analysis}
\usage{
\method{coef_profile}{mb_meta_analysis}(
  object,
  param_type = "fixed",
  include_constant = TRUE,
  conf_level = getOption("mb.conf_level", 0.95),
  estimate = getOption("mb.estimate", median),
  parallel = getOption("mb.parallel", FALSE),
  beep = getOption("mb.beep", TRUE),
  ...
)
}
\arguments{
\item{object}{The mb_meta_analysis object.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{estimate}{The function to use to calculate the estimate.}

\item{parallel}{A flag indicating whether to using parallel backend provided by foreach.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Not used.}
}
\value{
A tidy tibble.
}
\description{
Coef TMB Meta Analysis
}
