% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-profile.R
\name{coef_profile}
\alias{coef_profile}
\title{Coef Profile}
\usage{
coef_profile(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\description{
Gets coef confidence limits by likelihood profiling.
}
