% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm-comments.R
\name{comment_string}
\alias{comment_string}
\title{Comment String}
\usage{
comment_string(object, ...)
}
\arguments{
\item{object}{The mb code object.}

\item{...}{Unused.}
}
\value{
A string of the regular expression.
}
\description{
Returns the regular expression used to identify the start of a comment in an mb_code object.
}
