% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr.R
\name{converged.mb_analysis}
\alias{converged.mb_analysis}
\title{Is Converged}
\usage{
\method{converged}{mb_analysis}(x, rhat = getOption("mb.rhat", 1.1), ...)
}
\arguments{
\item{x}{Object to test rhat for}

\item{rhat}{A number specifying the rhat threshold.}

\item{...}{Unused}
}
\description{
Is Converged
}

