% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence.mb_analysis}
\alias{convergence.mb_analysis}
\title{Convergence}
\usage{
\method{convergence}{mb_analysis}(x, ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
convergence
}

