% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr.R
\name{derive_data}
\alias{derive_data}
\title{Derive Data}
\usage{
derive_data(object, new_data = data_set(object), new_expr = NULL,
  new_values = list(), term = "prediction", modify_new_data = NULL,
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep",
  FALSE), ...)
}
\arguments{
\item{object}{An object inheriting from class mb_analysis.}

\item{new_data}{The data frame to calculate the predictions for.}

\item{new_expr}{A string of R code specifying the predictive relationship.}

\item{new_values}{A named list of new or replacement values to pass to new_expr.}

\item{term}{A string of the term in new_expr.}

\item{modify_new_data}{A single argument function to modify new data (in list form) immediately prior to calculating new_expr.}

\item{parallel}{A flag indicating whether to do predictions using parallel backend provided by foreach.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\value{
A object of class mcmcr_data.
}
\description{
Calculate derived parameters.
}
