% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe}
\usage{
describe(x, latex = FALSE, quote = "", ...)
}
\arguments{
\item{x}{The model or analysis to describe.}

\item{latex}{A flag indicating whether to replace the parameters with their latex values.}

\item{quote}{A string specifying the text to place immediately before and after the parameters.}

\item{...}{Unused.}
}
\value{
A tibble with the columns parameter and description.
}
\description{
Describe the parameters in a model.
}
\details{
Requires that the model has a user-supplied description.
}

