% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.mb_analysis}
\alias{deviance.mb_analysis}
\title{Deviance}
\usage{
\method{deviance}{mb_analysis}(object, ...)
}
\arguments{
\item{object}{The MB analysis object.}

\item{...}{Unused.}
}
\value{
The analysis data set with the deviance values.
}
\description{
Extract deviance values for an MB analysis.
}
\details{
The new_expr in the model must include the term 'deviance'.
}
