% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-parameters.R
\name{drop_parameters}
\alias{drop_parameters}
\title{Drop Parameters}
\usage{
drop_parameters(x, parameters = character(0))
}
\arguments{
\item{x}{The object.}

\item{parameters}{A character vector of the parameters to drop.}
}
\value{
The updated object.
}
\description{
Drops named scalar fixed parameters from a model by fixing them at 0.
}

