% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-parameters.R
\name{drop_pars}
\alias{drop_pars}
\title{Drop pars}
\usage{
drop_pars(x, pars = character(0), ...)
}
\arguments{
\item{x}{The object.}

\item{pars}{A character vector of the pars to drop.}

\item{...}{Not used.}
}
\value{
The updated object.
}
\description{
Drops named scalar fixed pars from an object by fixing them at 0.
}
