% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate.R
\name{duration}
\alias{duration}
\title{Duration}
\usage{
duration(num = NULL, units = "seconds", ...)
}
\arguments{
\item{num}{the number of time units to include in the duration. From v1.6.0
\code{num} can also be a character vector that specifies durations in a
convenient shorthand format. All unambiguous name units and abbreviations
are supported. See examples.}

\item{units}{a character string that specifies the type of units that num
refers to. When \code{num} is character, this argument is ignored.}

\item{...}{a list of time units to be included in the duration and their
amounts. Seconds, minutes, hours, days, and weeks are supported.}
}
\description{
Duration
}
\seealso{
\code{\link[lubridate]{duration}}.
}

