% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embr-package.R
\docType{package}
\name{embr-package}
\alias{embr}
\alias{embr-package}
\title{embr: Model Builder Utility Functions and Virtual Classes}
\description{
Utility functions and virtual classes shared by model builder packages such as tmbr, jmbr and smbr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/embr}
  \item Report bugs at \url{https://github.com/poissonconsulting/embr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
