% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Get Estimates}
\usage{
estimates(object, terms = "fixed", scalar = FALSE, ...)
}
\arguments{
\item{object}{The object.}

\item{terms}{A string indicating the type of terms to get the estimates for.}

\item{scalar}{A flag indicating whether to only return scalar terms.}

\item{...}{Not used.}
}
\value{
The estimates as a named list.
}
\description{
Gets the estimates for an object as a named list.
}

