% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.lmb_analysis}
\alias{fitted.lmb_analysis}
\title{Fitted Values}
\usage{
\method{fitted}{lmb_analysis}(object, ...)
}
\arguments{
\item{object}{The LMB analysis object.}

\item{...}{Unused.}
}
\value{
The analysis data set with the fitted values.
}
\description{
Extract fitted values for a LMB analysis.
}
