% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.mb_analysis}
\alias{fitted.mb_analysis}
\title{Fitted Values}
\usage{
\method{fitted}{mb_analysis}(object, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{...}{Unused.}
}
\value{
The analysis data set with the fitted values.
}
\description{
Extract fitted values for a MB analysis.
}

