% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{IC.list}
\alias{IC.list}
\title{Information Criterion}
\usage{
\method{IC}{list}(object, n = NULL, ...)
}
\arguments{
\item{object}{The object to calculate it for.}

\item{n}{A count of the sample size.}

\item{...}{Unused.}
}
\description{
Calculate default information criterion.
}

