% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC}
\alias{IC}
\title{Information Criterion}
\usage{
IC(object, n = NULL, ...)
}
\arguments{
\item{object}{The object to calculate the IC for.}

\item{n}{An count of the sample size. If NULL it is assumed to be the number of rows in the data.}

\item{...}{Not used.}
}
\value{
The Information Criteron as a number.
}
\description{
Calculates Information Criterion for an analysis.
}
