% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ICw}
\alias{ICw}
\title{Information Criterion Weight}
\usage{
ICw(analyses, n = NULL)
}
\arguments{
\item{analyses}{The list of mb_analysis objects to calculate it for.}

\item{n}{A count of the sample size.}
}
\value{
A vector of the w_i values for each model.
}
\description{
Calculates w_i for a list of objects of class mb_analysis that return IC.
}

