% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.lmb_analysis}
\alias{is.lmb_analysis}
\title{Is a LMB Analysis}
\usage{
is.lmb_analysis(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'lmb_analysis'
}
