% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.lmb_model}
\alias{is.lmb_model}
\title{Is a LMB Model}
\usage{
is.lmb_model(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'lmb_model'
}
