% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mb_models}
\alias{is.mb_models}
\title{Is MB Models?}
\usage{
is.mb_models(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mb_models'
}
