% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.syntactic}
\alias{is.syntactic}
\title{Is syntactic}
\usage{
is.syntactic(x)
}
\arguments{
\item{x}{A character of possible variable names.}
}
\value{
A logical vector indicating whether a syntactically correct variable name.
}
\description{
Is syntactic
}
\examples{
is.syntactic(c("0", "x", "1x", "x y", "x1"))
}
