% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_bayesian}
\alias{is_bayesian}
\title{Is Bayesian}
\usage{
is_bayesian(x)
}
\arguments{
\item{x}{An analysis to test}
}
\value{
A flag indicating whether x is a Bayesian analysis.
}
\description{
Is Bayesian
}
