% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr.R
\name{is_converged.mb_analysis}
\alias{is_converged.mb_analysis}
\title{Is Converged}
\usage{
is_converged.mb_analysis(x, rhat = getOption("mb.rhat", 1.1), ...)
}
\arguments{
\item{x}{Object to test convergence for}

\item{rhat}{A number specifying the rhat threshold.}

\item{...}{Unused}
}
\description{
Is Converged
}

