% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-paradigm.R
\name{is_frequentist}
\alias{is_frequentist}
\title{Test if is frequentist}
\usage{
is_frequentist(x)
}
\arguments{
\item{x}{the object.}
}
\value{
A flag indicating whether frequentist
}
\description{
Test if is frequentist
}
