% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_namedlist}
\alias{is_namedlist}
\title{Is Named List}
\usage{
is_namedlist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag.
}
\description{
Is Named List
}
\examples{
is_namedlist(1)
is_namedlist(list())
is_namedlist(list(1))
is_namedlist(list(x = 1))
is_namedlist(list(x = list(y = 2)))
}
