% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nlist}
\alias{is_nlist}
\title{Is Named List}
\usage{
is_nlist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag.
}
\description{
Is Named List
}
\examples{
is_nlist(1)
is_nlist(list())
is_nlist(list(1))
is_nlist(list(x = 1))
is_nlist(list(x = list(y = 2)))
}

