% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-model.R
\name{load_model}
\alias{load_model}
\title{Load Model}
\usage{
load_model(x, quiet, ...)
}
\arguments{
\item{x}{The model to load.}

\item{quiet}{A flag indicating whether to suppress warnings and output.}

\item{...}{Additional arguments.}
}
\description{
Load Model
}
