% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik-draws.R
\name{log_lik_draws.mb_analysis}
\alias{log_lik_draws.mb_analysis}
\title{Extract log likelihood draws}
\usage{
\method{log_lik_draws}{mb_analysis}(x, joint = FALSE, log_lik_name = "log_lik", ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{joint}{A flag indicating whether to return the joint log
likelihood or array, default is FALSE.}

\item{log_lik_name}{A string of the name of the parameter corresponding to
the log likelihood, default is "log_lik".}

\item{...}{Unused.}
}
\value{
A draws_array object containing log_lik values.
}
\description{
Extract log likelihood from fitted model and return as a draws
object. Adapted from the `priorsense` package.
}
