% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-prior-draws.R
\name{log_prior_draws.mb_analysis}
\alias{log_prior_draws.mb_analysis}
\title{Extract log prior draws}
\usage{
\method{log_prior_draws}{mb_analysis}(x, joint = FALSE, log_prior_name = "lprior", ...)
}
\arguments{
\item{x}{Model fit or draws object.}

\item{joint}{Logical indicating whether to return the joint log
prior or array. Default is FALSE.}

\item{log_prior_name}{Name of parameter corresponding to log prior, default
is "lprior".}

\item{...}{Arguments passed to individual methods.}
}
\value{
A draws_array object containing log_prior values.
}
\description{
Extract log likelihood from fitted model and return as a draws
object. Adapted from the `priorsense` package.
}
