% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.lmb_analysis}
\alias{logLik.lmb_analysis}
\title{Log-Likelihood}
\usage{
\method{logLik}{lmb_analysis}(object, ...)
}
\arguments{
\item{object}{The lm_analysis object.}

\item{...}{unused.}
}
\description{
Log-Likelihood for a LM analysis.
}
