% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.mb_analysis}
\alias{logLik.mb_analysis}
\title{Log-Likelihood}
\usage{
\method{logLik}{mb_analysis}(object, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{...}{unused.}
}
\description{
Log-Likelihood for a MB analysis.
}

