% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicc.R
\name{mAICc}
\alias{mAICc}
\title{Akaike's Information Criteron}
\usage{
mAICc(object, ...)
}
\arguments{
\item{object}{The object to calculate the AIC for.}

\item{...}{Not used.}
}
\value{
The Akaike's Information Criteron as a number.
}
\description{
Calculates marginal AICc for an analysis.
}
\details{
By default the sample size is assumed to be the number of rows
in the analysis data set. To override manually set n to be an integer ie n = 10L.
To calculate mAIC (as opposed to mAICc) set n = Inf.
}

