% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-all-models.R
\name{make_all_models}
\alias{make_all_models}
\title{Make All Models}
\usage{
make_all_models(model, drops = list())
}
\arguments{
\item{model}{The full model.}

\item{drops}{A list of character vectors specifying possible drops.}
}
\value{
A list of objects inheriting from class mb_model.
}
\description{
Make All Models
}
