% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr-package.R
\docType{package}
\name{mbr-package}
\alias{mbr}
\alias{mbr-package}
\title{mbr: Model Builder Utility Functions and Virtual Classes}
\description{
Utility functions and virtual classes shared by model builder packages
    such as tmbr, jmbr and smbr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/mbr}
  \item Report bugs at \url{https://github.com/poissonconsulting/mbr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
