% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{mcmc_derive.mb_analyses}
\alias{mcmc_derive.mb_analyses}
\title{Derive}
\usage{
\method{mcmc_derive}{mb_analyses}(
  object,
  new_data = data_set(object),
  new_expr = NULL,
  new_values = list(),
  term = "prediction",
  modify_new_data = NULL,
  ref_data = FALSE,
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE),
  beep = getOption("mb.beep", FALSE),
  ...
)
}
\arguments{
\item{object}{An object inheriting from class mb_analysis.}

\item{new_data}{The data frame to calculate the predictions for.}

\item{new_expr}{A string of R code specifying the predictive relationship.}

\item{new_values}{A named list of new or replacement values to pass to new_expr.}

\item{term}{A string of the term in new_expr.}

\item{modify_new_data}{A single argument function to modify new data (in list form) immediately prior to calculating new_expr.}

\item{ref_data}{A flag or a data frame with 1 row indicating the reference values for calculating the effects size.}

\item{parallel}{A flag indicating whether to do predictions using parallel backend provided by foreach.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\value{
A object of class mcmcr.
}
\description{
Calculate derived parameters.
}
