% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive-data.R
\name{mcmc_derive_data}
\alias{mcmc_derive_data}
\title{Derive Data}
\usage{
mcmc_derive_data(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Not used.}
}
\description{
Calculate derived parameters.
}
