% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.lmb_code}
\alias{model.lmb_code}
\title{MB Model}
\usage{
\method{model}{lmb_code}(x, select_data = list(), ...)
}
\arguments{
\item{x}{An object inheriting from class mb_code.}

\item{select_data}{A named list specifying the columns to select and their associated classes.}

\item{...}{Unused arguments.}
}
\value{
An object inherting from class lmb_model.
}
\description{
Creates MB model.
}
\details{
For tmb models gen_inits must specify all the fixed parameters.
Missing random parameters are assigned the value 0.

For jmb models unspecified the initial values for each chain are drawn from the prior distributions.
}
\seealso{
\code{\link[datacheckr]{check_data2}} \code{\link[rescale]{rescale_c}}
}
