% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{MB Models}
\usage{
models(...)
}
\arguments{
\item{...}{Named objects.}
}
\value{
An object inheriting from class mb_models.
}
\description{
Creates an object inherting from class mb_models.
}
