% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-data.R
\name{modify_data}
\alias{modify_data}
\title{Modify Data}
\usage{
modify_data(data, model, numericize_factors = FALSE)
}
\arguments{
\item{data}{The data to modify.}

\item{model}{An object inheriting from class mb_model.}

\item{numericize_factors}{A flag indicating whether to convert factors to integer.}
}
\value{
The modified data in list form.
}
\description{
Modifies a data frame to the form it will be passed to the analysis code.
}
