% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply-sd-normal-priors.R
\name{multiply_sd_normal_priors}
\alias{multiply_sd_normal_priors}
\title{Multiply Standard Deviation of Normal Priors}
\usage{
multiply_sd_normal_priors(x, multiplier = 2, ...)
}
\arguments{
\item{x}{The object.}

\item{multiplier}{A double vector of the values to multiply by.}

\item{...}{Not used.}
}
\value{
The updated object.
}
\description{
Multiply Standard Deviation of Normal Priors
}
