% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr.R
\name{nchains.mb_analysis}
\alias{nchains.mb_analysis}
\title{Number of chains}
\usage{
\method{nchains}{mb_analysis}(x, ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{...}{Not used.}
}
\value{
A count of the number of chains.
}
\description{
nchains
}

