% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-analysis.R
\name{new_analysis}
\alias{new_analysis}
\title{MB Analysis}
\usage{
new_analysis(x, class)
}
\arguments{
\item{x}{A list.}

\item{class}{The class of the new object.}
}
\description{
Constucts a new analysis object
}
