% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-expr.R
\name{new_expr<-}
\alias{new_expr<-}
\title{new_expr set}
\usage{
new_expr(object) <- value
}
\arguments{
\item{object}{The object to set for.}

\item{value}{The new value of new expr.}
}
\value{
The modified object.
}
\description{
new_expr set
}
