% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngens.R
\name{ngens}
\alias{ngens}
\title{Total Number of MCMC simulations generated for each chain.}
\usage{
ngens(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A count.
}
\description{
Total Number of MCMC simulations generated for each chain.
}
