% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr.R
\name{niters.mb_analysis}
\alias{niters.mb_analysis}
\title{Number of iterations}
\usage{
\method{niters}{mb_analysis}(x, ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{...}{Not used.}
}
\value{
A count of the number of iterations.
}
\description{
niters
}
