% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmodels.R
\name{nmodels}
\alias{nmodels}
\title{Number Models}
\usage{
nmodels(x, ...)
}
\arguments{
\item{x}{the object.}

\item{...}{Named objects.}
}
\value{
An integer of the number of models
}
\description{
Number Models
}
