% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparams.R
\name{nparams}
\alias{nparams}
\title{Number of parameters}
\usage{
nparams(x, param_type = "fixed", scalar_only = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{param_type}{A string indicating the type of terms to get the names for.}

\item{scalar_only}{A flag indicating whether to only return scalar terms.}

\item{...}{Not used.}
}
\value{
A integer of the number of parameters.
}
\description{
Gets the parameter number for an object.
}
