% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsims.R
\name{nsims}
\alias{nsims}
\title{Total Number of MCMC simulations generated.}
\usage{
nsims(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A count.
}
\description{
Total Number of MCMC simulations generated.
}
