% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms.mb_analysis}
\alias{nterms.mb_analysis}
\title{Number of terms}
\usage{
\method{nterms}{mb_analysis}(x, fixed = TRUE, include_constant = TRUE, ...)
}
\arguments{
\item{x}{The object to get the nterms for}

\item{fixed}{A flag specifying whether fixed or random terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{...}{unused}
}
\description{
Number of terms
}

