% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\title{Number of fixed variable terms.}
\usage{
nterms(object, ...)
}
\arguments{
\item{object}{The object to calculate the number of terms for.}

\item{...}{Not used.}
}
\value{
A count of the number of terms.
}
\description{
Number of fixed variable terms.
}

