% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericize.R
\name{numericize_factors}
\alias{numericize_factors}
\title{Numericize Factors}
\usage{
numericize_factors(data)
}
\arguments{
\item{data}{A list.}
}
\value{
The list with factors converted to integers.
}
\description{
Numericize Factors
}
