% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Parameters}
\usage{
parameters(x, terms = "fixed", scalar = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{terms}{A string indicating the type of terms to get the names for.}

\item{scalar}{A flag indicating whether to only return scalar terms.}

\item{...}{Not used.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets the parameter names for an object.
}

