% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plapply}
\alias{plapply}
\title{plapply}
\usage{
plapply(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{optional arguments to \code{FUN}.}
}
\description{
plapply
}

