% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plapply_chunks}
\alias{plapply_chunks}
\title{Parallel lapply on chunks}
\usage{
plapply_chunks(X, FUN, .parallel = TRUE, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{.parallel}{A flag indicating whether to analyse chunks in parallel.}

\item{...}{optional arguments to \code{FUN}.}
}
\description{
Applys function FUN to entire list X or possibly on chunks of X in parallel.
}
