% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Residuals}
\usage{
plot_residuals(x, ...)
}
\arguments{
\item{x}{The object to plot the residuals for.}

\item{...}{Unused.}
}
\description{
Plot Residuals
}
