% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-predictive-check.R
\name{posterior_predictive_check}
\alias{posterior_predictive_check}
\title{Posterior Predictive Check}
\usage{
posterior_predictive_check(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A tibble of the checks.
}
\description{
Simulates
}
