% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale-sensitivity.R
\name{powerscale_sensitivity.mb_analysis}
\alias{powerscale_sensitivity.mb_analysis}
\title{Run Power-Scaling Sensitivity Analysis}
\usage{
\method{powerscale_sensitivity}{mb_analysis}(x, ...)
}
\arguments{
\item{x}{The MB analysis object}

\item{...}{Arguments passed to methods to
\code{\link[priorsense]{powerscale_sensitivity}}. Alternative functions for
`log_lik_fn` and `log_prior_fn` cannot be used.}
}
\value{
Table of sensitivity values for each specified variable.
}
\description{
Calculates the prior/likelihood sensitivity based on power-scaling
perturbations. This is done using importance sampling (and
optionally moment matching).
}
