% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check.mb_analysis}
\alias{pp_check.mb_analysis}
\title{Posterior predictive checks for an MB analysis}
\usage{
\method{pp_check}{mb_analysis}(object, y, nrep = 100L, plot = TRUE, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{y}{Character naming response to compare with posterior predictions.}

\item{nrep}{Integer number of posterior predictive replicates. Defaults to 100.}

\item{plot}{Should posterior predictions be plotted? Default is TRUE.}

\item{...}{unused.}
}
\value{
A list of \code{y} and \code{nrep} posterior predictions the same length as \code{y}.
}
\description{
Posterior predictive checks for an MB analysis
}
\seealso{
\link{pp_check}
}
