% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lmb_analysis}
\alias{predict.lmb_analysis}
\title{Predict}
\usage{
\method{predict}{lmb_analysis}(object, new_data = data_set(object),
  conf_level = 0.95, quiet = getOption("mb.quiet", TRUE),
  beep = getOption("mb.beep", FALSE), ...)
}
\arguments{
\item{object}{An object inheriting from class mb_analysis.}

\item{new_data}{The data frame to calculate the predictions for.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.#' @return A object of class mcmcr_data.}
}
\description{
Predict
}
