% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mb_analyses}
\alias{predict.mb_analyses}
\title{Predict from a \code{mb_analyses} object}
\usage{
\method{predict}{mb_analyses}(
  object,
  new_data = data_set(object),
  new_expr = NULL,
  new_values = list(),
  term = "prediction",
  conf_level = getOption("mb.conf_level", 0.95),
  modify_new_data = NULL,
  ref_data = FALSE,
  ref_fun2 = proportional_change2,
  new_expr_vec = getOption("mb.new_expr_vec", FALSE),
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE),
  ...
)
}
\arguments{
\item{object}{An object inheriting from class mb_analysis.}

\item{new_data}{A data frame to calculate the predictions for.
Pass \code{character(0)} to extract a scalar \code{term} from new_expr.}

\item{new_expr}{An R expression (e.g. \code{{ ... }}) or a character string of R
code specifying the predictive relationship. If \code{NULL}, uses the expression
set in \code{\link[=model]{model()}} and stored in the \code{mb_analysis} object.}

\item{new_values}{A named list of new or replacement values to pass to new_expr.}

\item{term}{A string of the term in new_expr.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{modify_new_data}{A single argument function to modify new data (in list form) immediately prior to calculating new_expr.}

\item{ref_data}{A flag or a data frame with 1 row indicating the reference values for calculating the effects size.
If \code{FALSE}, no reference applied. If \code{TRUE}, the reference is a 1 row data.frame with all variables held at reference value.}

\item{ref_fun2}{A function whose first argument takes a vector of two numbers and returns a scalar of a metric of the difference between them.}

\item{new_expr_vec}{A flag specifying whether to vectorize the new_expr code.}

\item{parallel}{A flag indicating whether to do predictions using parallel backend provided by foreach.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{...}{Additional arguments.}
}
\value{
A data frame with one row per row of \code{new_data}.
}
\description{
Returns IC-weighted model-averaged predictions across an \code{mb_analyses} list.
}
\seealso{
\code{\link[=predict.mb_analysis]{predict.mb_analysis()}} for full argument documentation and
examples.
}
