% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{process_data}
\alias{process_data}
\title{Process Data}
\usage{
process_data(data, data2, model)
}
\arguments{
\item{data}{The data to modify.}

\item{data2}{The base data.}

\item{model}{An object inheriting from class tmb_model.#'}
}
\value{
The modified data in list form.
}
\description{
Processes a data frame by selecting, centering and scaling data
}

