% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-coef.R
\name{profile_coef}
\alias{profile_coef}
\title{Profile Coef}
\usage{
profile_coef(object, conf_level = getOption("mb.conf_level", 0.95), ...)
}
\arguments{
\item{object}{The object.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Unused.}
}
\description{
Profile Coef
}
