% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2.mb_analysis}
\alias{R2.mb_analysis}
\title{R2}
\usage{
\method{R2}{mb_analysis}(object, response, marginal = TRUE,
  parallel = getOption("mb.parallel", FALSE), quiet = getOption("mb.quiet",
  TRUE), beep = getOption("mb.beep", FALSE), ...)
}
\arguments{
\item{object}{The object.}

\item{response}{A string specifying the column in the data corresponding to the response.}

\item{marginal}{A flag indicating whether to calculate the marginal or conditional R2 value.}

\item{parallel}{A flag indicating whether to do predictions using parallel backend provided by foreach.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Unused}
}
\value{
A number of the R2 value.
}
\description{
Gets the marginal or conditional R2 value on the original scale for an mb_analysis object.
}
