% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-effects.R
\name{random_effects}
\alias{random_effects}
\title{Random Effects}
\usage{
random_effects(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\value{
The random effects as a sorted named list.
}
\description{
Gets the random effects definitions for an object inheriting from class mb_model or mb_analysis.
}

