% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalyse.R
\name{reanalyse.mb_meta_analyses}
\alias{reanalyse.mb_meta_analyses}
\title{Reanalyse}
\usage{
\method{reanalyse}{mb_meta_analyses}(
  object,
  rhat = getOption("mb.rhat", 1.1),
  esr = getOption("mb.esr", 0.33),
  nreanalyses = getOption("mb.nreanalyses", 1L),
  duration = getOption("mb.duration", dhours(1)),
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE),
  glance = getOption("mb.glance", TRUE),
  beep = getOption("mb.beep", TRUE),
  ...
)
}
\arguments{
\item{object}{The object to reanalyse.}

\item{rhat}{A number specifying the rhat threshold.}

\item{esr}{A number specifying the effective sampling rate.}

\item{nreanalyses}{A count between 1 and 7 specifying the maximum number of reanalyses.}

\item{duration}{The maximum total time to spend on analysis/reanalysis.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{glance}{A flag indicating whether to print summary of model.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Unused arguments.}
}
\description{
Reanalyse
}
