% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalyse.R
\name{reanalyse}
\alias{reanalyse}
\title{Reanalyse}
\usage{
reanalyse(analysis, rhat = getOption("mb.rhat", 1.1),
  duration = getOption("mb.duration", dminutes(10)),
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE),
  glance = getOption("mb.glance", TRUE), beep = getOption("mb.beep", TRUE),
  ...)
}
\arguments{
\item{analysis}{An object inheriting from class mb_analysis or a list of such objects.}

\item{rhat}{A number specifying the rhat threshold.}

\item{duration}{The maximum total time to spend on analysis/reanalysis.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{glance}{A flag indicating whether to print summary of model.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Unused.}
}
\description{
Reanalyse an analysis.
}
