% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-latex.R
\name{replace_latex}
\alias{replace_latex}
\title{Replace LaTeX}
\usage{
replace_latex(term, latex)
}
\arguments{
\item{term}{A character vector of the terms.}

\item{latex}{A named character vector specifying the latex math code for the terms in coef.}
}
\value{
A character vector of the corresponding latex code.
}
\description{
Replaces terms with the values in latex.
}
\details{
Any indices are replaced with the corresponding latex codes for subscripted indices
ie b[1,3] is replaced by b_{1,3}.
}
\examples{
term <- c("a", "b", paste0("c", c("[1,1]", "[2,1,44]")))
replace_latex(term, c(a = "\\\\alpha", b = "\\\\beta_{\\\\lambda}", c = "\\\\beta Y"))
}

