% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample-residuals.R
\name{resample_residuals}
\alias{resample_residuals}
\title{Resample Residuals}
\usage{
resample_residuals(x)
}
\arguments{
\item{x}{The MB analysis object.}
}
\value{
An mcmc_data of the resampled residuals.
}
\description{
Requires that new_expr includes code like `residual <- res_bern(`
or `residual[i] <- res_norm(` without the `resample` argument.
}
