% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals.mb_analysis}
\alias{residuals.mb_analysis}
\title{Residuals}
\usage{
\method{residuals}{mb_analysis}(object, ...)
}
\arguments{
\item{object}{The MB analysis object.}

\item{...}{Unused.}
}
\value{
The analysis data set with the residual values.
}
\description{
Extract residual values for an MB analysis.
}
\details{
The new_expr in the model must include the term 'residual'.
}

