% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr.R
\name{rhat.mb_analysis}
\alias{rhat.mb_analysis}
\title{Rhat}
\usage{
\method{rhat}{mb_analysis}(x, ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
rhat
}
