% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{rm_comments}
\alias{rm_comments}
\title{Removes comments from the template strings(s) of an mb object or a character vector.}
\usage{
rm_comments(object, ...)
}
\arguments{
\item{object}{The mb object or character vector.}

\item{...}{Unused.}
}
\value{
The mb object or character vector without comments in its template string(s).
}
\description{
Removes comments from the template strings(s) of an mb object or a character vector.
}
