% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{sample_size}
\alias{sample_size}
\title{Sample Size}
\usage{
sample_size(object, ...)
}
\arguments{
\item{object}{The object to calculate the sample size for.}

\item{...}{Not used.}
}
\value{
A count of the sample size.
}
\description{
Get sample size.
}
