% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scalar_nlist}
\alias{scalar_nlist}
\title{Scalar Named List}
\usage{
scalar_nlist(x)
}
\arguments{
\item{x}{The named list to sort.}
}
\value{
The sorted named list.
}
\description{
Filters a named list so only scalar elements remain.wiby its names.
}
\examples{
scalar_nlist(list(y = 2, x = 1, a = c(10,1)))
}

