% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-priors-by.R
\name{sd_priors_by}
\alias{sd_priors_by}
\alias{sd_priors_by.mb_model}
\alias{sd_priors_by.mb_analysis}
\title{Multiply Standard Deviation of Priors By}
\usage{
sd_priors_by(x, by = 10, ...)

\method{sd_priors_by}{mb_model}(x, by = 10, ...)

\method{sd_priors_by}{mb_analysis}(x, by = 10,
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE), glance = getOption("mb.glance",
  TRUE), beep = getOption("mb.beep", TRUE), ...)
}
\arguments{
\item{x}{The object.}

\item{by}{A double scalar of the multiplier.}

\item{...}{Not used.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{glance}{A flag indicating whether to print a model summary.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}
}
\value{
The updated object.
}
\description{
Multiply Standard Deviation of Priors By
}
\section{Methods (by class)}{
\itemize{
\item \code{mb_model}: Multiply Standard Deviation of Priors for an MB model

\item \code{mb_analysis}: Multiply Standard Deviation of Priors for an MB analysis
}}

