% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-data.R
\name{select_rescale_data}
\alias{select_rescale_data}
\title{Select and Rescale Data}
\usage{
select_rescale_data(data, model, data2 = data)
}
\arguments{
\item{data}{The data to modify.}

\item{model}{An object inheriting from class mb_model.}

\item{data2}{The base data.}
}
\value{
The modified data in list form.
}
\description{
Selects and rescales data.
}
