% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Summarize model sensitivity}
\usage{
sensitivity(x, by = "term", param_type = "all", ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{by}{A string indicating whether to determine by "term", "parameter", or
"all".}

\item{param_type}{A string specifying which parameters to include: 'fixed',
'random', 'derived', 'primary', or 'all'.}

\item{...}{Arguments passed to \code{\link[=add_sensitivity]{add_sensitivity()}}.}
}
\value{
A dataframe summarizing the sensitivity of the analysis object.
}
\description{
Summarize model sensitivity
}
