% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-analysis-mode.R
\name{set_analysis_mode}
\alias{set_analysis_mode}
\title{Set Analysis Mode}
\usage{
set_analysis_mode(mode = "report")
}
\arguments{
\item{mode}{A string of the analysis mode.}
}
\value{
The old options.
}
\description{
Sets analysis mode.
}
\details{
The possible modes are as follows:
\describe{
\item{'debug'}{To rapidly identify problems with a model definition.}
\item{'quick'}{To quickly test code runs.}
\item{'report'}{To produce results for a report.}
\item{'paper'}{To produce results for a peer-reviewed paper.}
\item{'check'}{To run when checking a package.}
\item{'reset'}{To reset all the options to NULL so that they are the default values for each function call.}
}

In each case the mode is a unique combination of the following package options
\describe{
\item{mb.nchains}{A count of the number of chains.}
\item{mb.niters}{A count of the number of simulations to save per chain.}
\item{mb.nthin}{A count of the thining interval.}
\item{mb.parallel}{A flag indicating whether to perform the analysis in parallel.}
\item{mb.quiet}{A flag indicating whether to disable tracing information.}
\item{mb.beep}{A flag indicating whether to beep on completion of the analysis.}
\item{mb.glance}{A flag indicating whether to print a model summary.}
\item{mb.nreanalyses}{A count specifying the maximum number of reanalyses.}
\item{mb.rhat}{A number specifying the rhat threshold.}
\item{mb.esr}{A number specifying the minimum effective sampling rate.}
\item{mb.prop_divergent}{A number specifying the maximum proportion of divergent transitions.}
\item{mb.duration}{The maximum total time to spend on analysis and reanalysis.}
\item{mb.conf_level}{A number specifying the confidence level.}
}
}
\examples{
\dontrun{
set_analysis_mode("reset")
}
}
