% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-residuals.R
\name{simulate_residuals}
\alias{simulate_residuals}
\title{Simulate Residuals}
\usage{
simulate_residuals(x, type = NULL)
}
\arguments{
\item{x}{The MB analysis object.}

\item{type}{A string of the residual type.}
}
\value{
An mcmc_data of the simulated residuals.
}
\description{
Requires that new_expr includes \verb{residual <- res_bern(}
or \verb{residual[i] <- res_norm(}.
}
\seealso{
extras::res_binom
}
