% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by.R
\name{sort_by_ic}
\alias{sort_by_ic}
\title{Sort Analyses by IC}
\usage{
sort_by_ic(x, ...)
}
\arguments{
\item{x}{the object.}

\item{...}{Unused}
}
\value{
The sorted object
}
\description{
Sort Analyses by IC
}
