% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_nlist}
\alias{sort_nlist}
\title{Sort Named List}
\usage{
sort_nlist(x)
}
\arguments{
\item{x}{The named list to sort.}
}
\value{
The sorted named list.
}
\description{
Sorts a named list by its names.
}
\examples{
sort_nlist(list(y = 2, x = 1, a = 10))
}

