% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template<-}
\alias{template<-}
\title{Set Template}
\usage{
template(object) <- value
}
\arguments{
\item{object}{The object.}

\item{value}{A string of the new template}
}
\description{
Sets the template for an object.
}
