% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Template}
\usage{
template(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Additional arguments.}
}
\value{
The template model code as a string.
}
\description{
Gets the template string for an object.
}
