% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terms.mb_analysis}
\alias{terms.mb_analysis}
\title{Terms}
\usage{
\method{terms}{mb_analysis}(x, fixed = TRUE, include_constant = TRUE, ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{fixed}{A flag specifying whether fixed or random terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{...}{Not used.}
}
\description{
terms
}

