% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-normal-priors.R
\name{update_normal_priors}
\alias{update_normal_priors}
\title{Update Normal Priors}
\usage{
update_normal_priors(object, multiplier = 2, ...)
}
\arguments{
\item{object}{The object to adjust the priors for.}

\item{multiplier}{A positive number indicating the adjustment to the standard deviation as a multiplier.}

\item{...}{Additional arguments.}
}
\description{
Adjusts standard deviation of simple normal priors in object by multiplier.
}
\details{
A simple normal distribution has a mean of 0 and a numeric standard deviation
(without truncation).
}
