% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayize.R
\name{vectorize}
\alias{vectorize}
\title{Vectorize}
\usage{
vectorize(array, factors)
}
\arguments{
\item{array}{The matrix or array.}

\item{factors}{A named list of two or more factors.}
}
\value{
A vector.
}
\description{
Converts a matrix or array into a vector based on two or more factors.
}
\details{
The function can be used in new_expr code to convert arrays or matrices
to vectors that line up with factors for exporting.
}

