% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-sensitivity.R
\name{add_sensitivity}
\alias{add_sensitivity}
\title{Add sensitivity data frame to an analysis object}
\usage{
add_sensitivity(x, new_expr = NULL, replace = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{new_expr}{A string of R code specifying the predictive relationship.}

\item{replace}{A flag specifying whether to replace an existing sensitivity
data frame.}

\item{...}{Arguments passed to \code{\link[=powerscale_sensitivity]{powerscale_sensitivity()}}.}
}
\value{
The analysis object with an added \code{sensitivity} data frame.
}
\description{
Computes power-scaling sensitivity and stores the result as \code{x$sensitivity}.
Subsequent calls return \code{x} unchanged unless \code{replace = TRUE}.
}
