% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse.character}
\alias{analyse.character}
\title{Analyse}
\usage{
\method{analyse}{character}(x, data, select_data = list(),
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE),
  glance = getOption("mb.glance", TRUE), beep = getOption("mb.beep", TRUE),
  ...)
}
\arguments{
\item{x}{An object inheriting from class mb_model or a list of such objects.}

\item{data}{The data frame to analyse.}

\item{select_data}{A named list specifying the columns to select and their associated classes and values as well as transformations and scaling options.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{glance}{A flag indicating whether to print a model summary.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Analyse
}
