% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse.mb_model}
\alias{analyse.mb_model}
\title{Analyse}
\usage{
\method{analyse}{mb_model}(x, data, seed = NA,
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE),
  glance = getOption("mb.glance", TRUE), beep = getOption("mb.beep", TRUE),
  ...)
}
\arguments{
\item{x}{An object inheriting from class mb_model or a list of such objects.}

\item{data}{The data frame to analyse.}

\item{seed}{An integer of the random seed. If \code{as.integer} produces \code{NA} the seed is generated randomly.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{glance}{A flag indicating whether to print a model summary.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Analyse
}
