% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse.mb_model}
\alias{analyse.mb_model}
\title{Analyse}
\usage{
\method{analyse}{mb_model}(x, data, nchains = getOption("mb.nchains",
  3L), niters = getOption("mb.niters", 1000L),
  nthin = getOption("mb.thin", NULL),
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE), glance = getOption("mb.glance",
  TRUE), beep = getOption("mb.beep", TRUE), ...)
}
\arguments{
\item{x}{An object inheriting from class mb_model or a list of such objects.}

\item{data}{The data frame to analyse.}

\item{nchains}{A count of the number of chains.}

\item{niters}{A count of the number of simulations to save per chain.}

\item{nthin}{A count of the thining interval or NULL (in which case taken from model).}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{glance}{A flag indicating whether to print a model summary.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Analyse
}
