% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse.mb_models}
\alias{analyse.mb_models}
\title{Analyse}
\usage{
\method{analyse}{mb_models}(
  x,
  data,
  nchains = getOption("mb.nchains", 3L),
  niters = getOption("mb.niters", 1000L),
  nthin = getOption("mb.thin", NULL),
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE),
  glance = getOption("mb.glance", TRUE),
  beep = getOption("mb.beep", TRUE),
  stan_engine = getOption("mb.stan_engine", character(0)),
  ...
)
}
\arguments{
\item{x}{An object inheriting from class mb_model or a list of such objects.}

\item{data}{The data frame to analyse.}

\item{nchains}{A count of the number of chains (default: 3).}

\item{niters}{A count of the number of iterations to save per chain (default: 1000).}

\item{nthin}{A count of the thinning interval.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quiet}{A flag indicating whether to disable messages and warnings, including sampling progress.}

\item{glance}{A flag indicating whether to print a model summary.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{stan_engine}{A string specifying the Stan engine to use:
\itemize{
\item \code{"rstan"} for MCMC sampling via \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (default).
\item \code{"cmdstan-mcmc"} for MCMC sampling via \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}
\item \code{"cmdstan-pathfinder"} for pathfinder estimation via \code{\link[cmdstanr:model-method-pathfinder]{cmdstanr::pathfinder()}}
\item \code{"cmdstan-optimize"} for optimization via \code{\link[cmdstanr:model-method-optimize]{cmdstanr::optimize()}}
\item \code{"cmdstan-laplace"} for Laplace approximation via \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}}
}}

\item{...}{Additional arguments.}
}
\description{
Analyse
}
