% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Analyse}
\usage{
analyse(model, data, beep = getOption("mb.beep", TRUE),
  debug = getOption("mb.debug", FALSE), ...)
}
\arguments{
\item{model}{The model to use for the analysis.}

\item{data}{The data frame to analyse.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{debug}{A flag indicating whether to run in debug mode.}

\item{...}{Additional arguments.}
}
\description{
Analyse a data set and model.
}

