% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Analyse}
\usage{
analyse(model, data, drop, quick = getOption("mb.quick", FALSE),
  quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep", TRUE),
  ...)
}
\arguments{
\item{model}{The model to use for the analysis.}

\item{data}{The data frame to analyse.}

\item{drop}{A character vector of scalar parameters to drop (fix at 0).}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Analyse a data set and model.
}

