% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse1}
\alias{analyse1}
\title{Analyse}
\usage{
analyse1(model, data, loaded, nchains, niters, nthin, quiet, glance,
  parallel, ...)
}
\arguments{
\item{model}{The mb_model to analyse.}

\item{data}{The data.}

\item{loaded}{The loaded model.}

\item{nchains}{chains.}

\item{niters}{iters}

\item{nthin}{thin}

\item{quiet}{quiet}

\item{glance}{glance}

\item{parallel}{parallel}

\item{...}{Additional arguments.}
}
\description{
Analyse
}
