% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse1}
\alias{analyse1}
\title{Analyse}
\usage{
analyse1(
  model,
  data,
  loaded,
  nchains,
  niters,
  nthin,
  quiet,
  glance,
  parallel,
  seed,
  niters_warmup,
  ...
)
}
\arguments{
\item{model}{The mb_model to analyse.}

\item{data}{The data.}

\item{loaded}{The loaded model.}

\item{nchains}{A count of the number of chains (default: 3).}

\item{niters}{A count of the number of iterations to save per chain (default: 1000).}

\item{nthin}{A count of the thinning interval.}

\item{quiet}{A flag indicating whether to disable messages and warnings, including sampling progress.}

\item{glance}{A flag indicating whether to print a model summary.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{seed}{A positive whole number specifying the seed to use. The default is random. This is currently only implemented for Stan models.}

\item{niters_warmup}{A count of the number of warmup iterations. The default is to use the same number of iterations as \code{niters}. This is currently only implemented for Stan models.}

\item{...}{Additional arguments.}
}
\description{
Analyse
}
