% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-data.R
\name{analyse_data}
\alias{analyse_data}
\title{Analyse Data}
\usage{
analyse_data(data, model, beep = getOption("mb.beep", TRUE),
  debug = getOption("mb.debug", FALSE), ...)
}
\arguments{
\item{data}{The data frame to analyse.}

\item{model}{The model to use for the analysis.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{debug}{A flag indicating whether to run in debug mode.}

\item{...}{Additional arguments.}
}
\description{
Analyse data using an object inheriting from class mb_model.
}

