% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayize.R
\name{arrayize}
\alias{arrayize}
\title{Arrayize}
\usage{
arrayize(list, vector, factors)
}
\arguments{
\item{list}{The named list with the vector and factors.}

\item{vector}{A string specifying the vector.}

\item{factors}{A character vector specifying two or more factors}
}
\value{
The modified list.
}
\description{
Converts a vector in a named list into a matrix or array based on two or more factors.
}
\details{
The function can be used in modify_data functions to convert vectors to matrices or
arrays.
}

