% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards.R
\name{backwards}
\alias{backwards}
\title{Backwards}
\usage{
backwards(model, data, drops = list(), conf_level = 0.95,
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep",
  TRUE), ...)
}
\arguments{
\item{model}{The model to use for the analysis.}

\item{data}{The data frame to analyse.}

\item{drops}{A list of character vectors specifying the scalar parameters to consider.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\value{
A list of the analyses.
}
\description{
Perform backwards step-wise regression on a model.
Returns a list of the analysis at each step starting with the full model.
}
\details{
drop is a list of character vectors specifying the scalar parameters to possibly drop.
If a list element consists of two or more strings then the earlier strings
are only available to drop after the later strings have been eliminated.
This allows polynomial dependencies to be respected.
}

