% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.list}
\alias{coef.list}
\title{Coef TMB Analyses}
\usage{
\method{coef}{list}(object, terms = "fixed", scalar_only = FALSE,
  constant_included = TRUE, conf_level = 0.95, n = NULL, ...)
}
\arguments{
\item{object}{The list of tmb_analysis objects.}

\item{terms}{A string of the type of terms to get the coefficients for.}

\item{scalar_only}{A flag indicating whether to only return scalar terms.}

\item{constant_included}{A flag indicating whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{n}{A count of the sample size.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coeffcient terms with the model averaged estimate, the
Akaike's weight and the proportion of models including the term.
}
\description{
Coefficients for a list of MB analyses averaged by IC weights.
}

