% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.list}
\alias{coef.list}
\title{Coef TMB Analyses}
\usage{
\method{coef}{list}(object, param_type = "fixed", include_constant = TRUE,
  conf_level = 0.95, n = NULL, ...)
}
\arguments{
\item{object}{The list of tmb_analysis objects.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{n}{A count of the sample size.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coeffcient terms with the model averaged estimate, the
Akaike's weight and the proportion of models including the term.
}
\description{
Coefficients for fixed parameters from a list of ML based
MB analyses averaged by AICc weights.
}

