% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mb_analysis}
\alias{coef.mb_analysis}
\title{Coef JAGS Analysis}
\usage{
\method{coef}{mb_analysis}(object, terms = "fixed", scalar_only = FALSE,
  constant_included = TRUE, conf_level = 0.95, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{terms}{A string of the type of terms to get the coefficients for.}

\item{scalar_only}{A flag indicating whether to only return scalar terms.}

\item{constant_included}{A flag indicating whether to include constant terms.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
Coefficients for a JAGS analysis.
}
\details{
Permitted values for terms are 'fixed' and 'random'.

The \code{statistic} is the z value.
}

