% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-set.R
\name{data_set}
\alias{data_set}
\title{Data Set}
\usage{
data_set(analysis, modify = FALSE, ...)
}
\arguments{
\item{analysis}{The object.}

\item{modify}{A flag indicating whether to modify the data.}

\item{...}{Unused.}
}
\value{
The data set as a tibble.
}
\description{
Gets the data set for an object inheriting from class mb_analysis.
}
