% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-set.R
\name{data_set}
\alias{data_set}
\title{Data Set}
\usage{
data_set(
  x,
  modify = FALSE,
  numericize_factors = FALSE,
  marginalize_random_effects = FALSE,
  ...
)
}
\arguments{
\item{x}{The object.}

\item{modify}{A flag indicating whether to modify the data.}

\item{numericize_factors}{A flag indicating whether to convert factors to integers if modifying the data.}

\item{marginalize_random_effects}{A flag indicating whether to set each factor in one or more random effects at its first level.}

\item{...}{Unused.}
}
\value{
The data set as a tibble.
}
\description{
Gets the data set for an object inheriting from class mb_analysis.
}
