% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embr-package.R
\docType{package}
\name{embr-package}
\alias{embr}
\alias{embr-package}
\title{embr: Model Builder Utility Functions and Virtual Classes}
\description{
Utility functions and virtual classes shared by model builder packages such as tmbr, jmbr and smbr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/embr}
  \item Report bugs at \url{https://github.com/poissonconsulting/embr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

Other contributors:
\itemize{
  \item Nicole Hill \email{nicole@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7623-2153}{ORCID}) [contributor]
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
  \item Nadine Hussein \email{nadine@poissonconsulting.ca} (\href{https://orcid.org/0000-0003-4470-8361}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
