% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.mb_analysis}
\alias{estimates.mb_analysis}
\title{Estimates}
\usage{
\method{estimates}{mb_analysis}(object, terms = "fixed",
  scalar_only = FALSE, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{terms}{A string of the type of terms to get the coefficients for.}

\item{scalar_only}{A flag indicating whether to only return scalar terms.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
estimates
}

