% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.mb_analysis}
\alias{estimates.mb_analysis}
\title{Estimates}
\usage{
\method{estimates}{mb_analysis}(x, param_type = "fixed", ...)
}
\arguments{
\item{x}{An object.}

\item{param_type}{A string indicating the type of terms to get the names for.}

\item{...}{Other arguments passed to methods.}
}
\value{
A list of uniquely named numeric objects.
}
\description{
Calculates the estimates for an MCMC object.
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{bind_iterations}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{split_chains}()}
}
