% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.mb_analysis}
\alias{estimates.mb_analysis}
\title{Estimates}
\usage{
\method{estimates}{mb_analysis}(object, fixed = TRUE, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{fixed}{A flag specifying whether fixed or random terms.}

\item{...}{Not used.}
}
\description{
estimates
}

