% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.mb_analysis}
\alias{estimates.mb_analysis}
\title{Estimates}
\usage{
\method{estimates}{mb_analysis}(x, param_type = "fixed", ...)
}
\arguments{
\item{x}{An object.}

\item{param_type}{A string indicating the type of terms to get the names for.}

\item{...}{Other arguments passed to methods.}
}
\value{
An \code{\link[nlist:nlist]{nlist::nlist()}} object.
}
\description{
Calculates the estimates for an MCMC object.
}
\examples{
estimates.foobar <- function(x, ...) {
 NotYetImplemented()
 # replace with code to get estimates for an object of class 'foobar'
}
}
\seealso{
Other {MCMC manipulations}: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{fill_all}()},
\code{\link[universals]{fill_na}()},
\code{\link[universals]{split_chains}()}
}
