% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Get Estimates}
\usage{
estimates(object, terms = "fixed", ...)
}
\arguments{
\item{object}{The object.}

\item{terms}{A string indicating the parameters to get the estimates for.}

\item{...}{Additional arguments.}
}
\value{
The estimates as a named list.
}
\description{
Gets the estimates for an object as a named list.
}

