% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{IC}
\alias{IC}
\title{Information Criterion}
\usage{
IC(object, ...)
}
\arguments{
\item{object}{The object to calculate it for.}

\item{...}{Not used.}
}
\description{
Calculate default information criterion.
}

