% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-prior-draws.R
\name{log_prior_draws.mb_analysis}
\alias{log_prior_draws.mb_analysis}
\title{Extract log prior draws}
\usage{
\method{log_prior_draws}{mb_analysis}(x, joint = FALSE, log_prior_name = "lprior", ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{joint}{A flag indicating whether to return the joint log
likelihood or array, default is FALSE.}

\item{...}{Unused.}

\item{log_lik_prior}{A string of the name of the parameter corresponding to
the log prior, default is "lprior".}
}
\value{
A draws_array object containing log_prior values.
}
\description{
Extract log likelihood from fitted model and return as a draws
object. Adapted from the `priorsense` package.
}
