% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.mb_code}
\alias{model.mb_code}
\title{MB Model}
\usage{
\method{model}{mb_code}(x, gen_inits, random_effects = list(),
  select_data = list(), center = character(0), scale = character(0),
  modify_data = identity, niters = 3L, new_expr = character(0),
  modify_new_data = identity, drops = list(), latex = character(0),
  description = character(0), ...)
}
\arguments{
\item{x}{An object inheriting from class mb_code.}

\item{gen_inits}{A single argument function taking the modified data and
returning a named list of initial values for all fixed parameters.
Missing random parameters are assigned the value 0.}

\item{random_effects}{A named list specifying the random effects and the associated factors.}

\item{select_data}{A named list specifying the columns to select and their associated classes and values.}

\item{center}{A character vector of the columns to center.}

\item{scale}{A character vector of the columns to scale (after centering).}

\item{modify_data}{A single argument function to modify the data (in list form) immediately prior to the analysis.}

\item{niters}{A count between 3 and 6 specifying the order of the number of iterations.}

\item{new_expr}{A string of R code specifying the predictive relationships.}

\item{modify_new_data}{A single argument function to modify new data (in list form) immediately prior to calculating new_expr.}

\item{drops}{A list of character vector of possible scalar parameters to drop (fix at 0).}

\item{latex}{A named character vector of the latex math code for each parameter.}

\item{description}{A named character vector describing each parameter.}

\item{...}{Unused arguments.}
}
\value{
An object inherting from class mb_model.
}
\description{
Creates TMB model.
}
\seealso{
\code{\link[datacheckr]{check_data}} \code{\link[rescale]{rescale}}
}

