% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{MB Model}
\usage{
model(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
An object inheriting from class mb_model.
}
\description{
Creates or retrieves an object inherting from class mb_model.
}

