% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{pars to monitor}
\usage{
monitor(object, param_type = "all")
}
\arguments{
\item{object}{An mb model object to get the pars for.}

\item{param_type}{A string specifying the type of pars to get.}
}
\value{
A character vector of the pars to monitor.
}
\description{
pars to monitor
}
