% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Parameters to monitor}
\usage{
monitor(object, param_type = "all", scalar_only = FALSE)
}
\arguments{
\item{object}{An mb model object to get the parameters for.}

\item{param_type}{A string specifying the type of parameters to get.}

\item{scalar_only}{A flag indicating whether to only get scalar parameters.}
}
\value{
A character vector of the parameters to monitor.
}
\description{
Parameters to monitor
}
