% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms.mb_analysis}
\alias{nterms.mb_analysis}
\title{Number of terms}
\usage{
\method{nterms}{mb_analysis}(x, param_type = "fixed",
  include_constant = TRUE, ...)
}
\arguments{
\item{x}{The object to get the nterms for}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{...}{unused}
}
\description{
Number of terms
}

