% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{param_type}{A string indicating the type of terms to get the names for.}

\item{type}{A string of the residual type.}

\item{stan_engine}{A string specifying the Stan engine in CmdStan to use.
Currently only "cmdstan-mcmc" is implemented.
If any other value is specified, the analysis will default to using 'rstan' via the 'smbr' package.}

\item{...}{Unused.}
}
\description{
Parameter Descriptions
}
