% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{param_type}{A string indicating the type of terms to get the names for.}

\item{type}{A string of the residual type.}

\item{stan_engine}{A string specifying the Stan engine to use:
\itemize{
  \item \code{"cmdstan-mcmc"} - Uses CmdStanR for MCMC sampling
  \item \code{character(0)} (default) or any other value - Uses RStan
}}

\item{...}{Unused.}
}
\description{
Parameter Descriptions
}
