% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mb_analysis}
\alias{plot.mb_analysis}
\title{Plot Analysis}
\usage{
\method{plot}{mb_analysis}(x, param_type = "fixed", ...)
}
\arguments{
\item{x}{The analysis object to plot}

\item{param_type}{A string indicating the type of terms to get the names for.}

\item{...}{Unused.}
}
\description{
Plot Analysis
}
