% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mb_analysis}
\alias{plot.mb_analysis}
\title{Plot Analysis}
\usage{
\method{plot}{mb_analysis}(x, fixed = TRUE, ...)
}
\arguments{
\item{fixed}{A flag specifying whether fixed or random terms.}
}
\description{
Plot Analysis
}

