% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-predictive-check.R
\name{posterior_predictive_check.mb_analysis}
\alias{posterior_predictive_check.mb_analysis}
\title{Posterior Predictive Check}
\usage{
\method{posterior_predictive_check}{mb_analysis}(x, zeros = TRUE, ...)
}
\arguments{
\item{x}{The MB analysis object.}

\item{zeros}{A flag specifying whether to perform a posterior predictive check
on the number of zeros in the data.}

\item{...}{Unused.}
}
\value{
A tibble of the checks.
}
\description{
Posterior Predictive Check
}
