% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check.mb_analysis}
\alias{pp_check.mb_analysis}
\title{Posterior predictive checks for an MB analysis}
\usage{
\method{pp_check}{mb_analysis}(analysis, y, nrep = 100L, plot = TRUE, ...)
}
\arguments{
\item{analysis}{The mb_analysis object.}

\item{y}{Character naming response to compare with posterior predictions.}

\item{nrep}{Integer number of posterior predictive replicates. Defaults to 100.}

\item{plot}{Should posterior predictions be plotted? Default is TRUE.}

\item{...}{unused.}
}
\value{
A list of \code{y} and \code{nrep} posterior predictions the same length as \code{y}.
}
\description{
S3 generic with simple default method. The intent is to provide a generic so
authors of other \R packages who wish to provide interfaces to the functions
in \pkg{bayesplot} will be encouraged to include \code{pp_check} methods in
their package, preserving the same naming conventions for posterior
predictive checking across many \R packages for Bayesian inference. This is
for the convenience of both users and developers. See the \strong{Details}
and \strong{Examples} sections, below, and the package vignettes for examples
of defining \code{pp_check} methods.
}
\seealso{
pp_check
}
