% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mb_analysis}
\alias{predict.mb_analysis}
\title{Predict}
\usage{
\method{predict}{mb_analysis}(object, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{...}{Unused.}
}
\value{
The new data set with the predictions.
}
\description{
Calculate predictions.
}

