% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-data.R
\name{predict_data}
\alias{predict_data}
\title{Predict Data}
\usage{
predict_data(data, analysis, new_expr = NULL, new_values = list(),
  term = "prediction", conf_int = FALSE, conf_level = 0.95,
  modify_new_data = NULL, parallel = getOption("mb.parallel", FALSE),
  quick = getOption("mb.quick", FALSE), quiet = getOption("mb.quiet", TRUE),
  beep = getOption("mb.beep", FALSE), ...)
}
\arguments{
\item{data}{The data frame to calculate the predictions for.}

\item{analysis}{An object inheriting from class mb_analysis.}

\item{new_expr}{A string of R code specifying the predictive relationship.}

\item{new_values}{A named list of new or replacement values to pass to new_expr.}

\item{term}{A string of the term in new_expr.}

\item{conf_int}{A flag indicating whether to calculate confidence intervals.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{modify_new_data}{A single argument function to modify new data (in list form) immediately prior to calculating new_expr.}

\item{parallel}{A flag indicating to do predictions using parallel backend provided by foreach.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Calculate predictions.
}

