% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-data.R
\name{predict_data}
\alias{predict_data}
\title{Predict Data}
\usage{
predict_data(data, analysis, new_expr = NULL, term = "prediction",
  conf_int = FALSE, conf_level = 0.95, ...)
}
\arguments{
\item{data}{The data frame to calculate the predictions for.}

\item{analysis}{An object inheriting from class mb_analysis.}

\item{new_expr}{A string of R code specifying the predictive relationship.}

\item{term}{A string of the term in new_expr.}

\item{conf_int}{A flag indicating whether to calculate confidence intervals.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Additional arguments.}
}
\description{
Calculate predictions.
}

