% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalyse.R
\name{reanalyse}
\alias{reanalyse}
\title{Reanalyse}
\usage{
reanalyse(analysis, quick = getOption("mb.quick", FALSE),
  quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep", TRUE),
  ...)
}
\arguments{
\item{analysis}{An object inheriting from class mb_analysis or a list of such objects.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Reanalyse an analysis.
}

