% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalyse.R
\name{reanalyse}
\alias{reanalyse}
\title{Reanalyse}
\usage{
reanalyse(analysis, duration = getOption("mb.duration", duration(1, "hours")),
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep",
  TRUE), ...)
}
\arguments{
\item{analysis}{An object inheriting from class mb_analysis or a list of such objects.}

\item{duration}{The target maximum amount of time to spend on analysis.}

\item{parallel}{A flag indicating whether to perform the analysis in parallel if possible.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\description{
Reanalyse an analysis.
}

