% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalyse.R
\name{reanalyse}
\alias{reanalyse}
\title{Reanalyse}
\usage{
reanalyse(object, ...)
}
\arguments{
\item{object}{The object to reanalyse.}

\item{...}{Additional arguments.}
}
\description{
Reanalyse an analysis.
}
