% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{estimates}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{universals}{\code{\link[universals]{estimates}}}
}}

