% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/coef.R, R/derive-data.R,
%   R/estimates.R, R/extras.R, R/glance.R, R/log-lik-draws.R,
%   R/log-prior-draws.R, R/logLik.R, R/powerscale-sensitivity.R, R/predict.R,
%   R/terms.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{coef}
\alias{mcmc_derive_data}
\alias{estimates}
\alias{odds_ratio2}
\alias{log_odds_ratio2}
\alias{proportional_change2}
\alias{proportional_difference2}
\alias{glance}
\alias{log_lik_draws}
\alias{log_prior_draws}
\alias{logLik}
\alias{powerscale_sensitivity}
\alias{fitted}
\alias{predict}
\alias{terms}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{extras}{\code{\link[extras]{log_odds_ratio2}}, \code{\link[extras]{odds_ratio2}}, \code{\link[extras]{proportional_change2}}, \code{\link[extras]{proportional_difference2}}}

  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{mcmcdata}{\code{\link[mcmcdata]{mcmc_derive_data}}}

  \item{priorsense}{\code{\link[priorsense]{log_lik_draws}}, \code{\link[priorsense]{log_prior_draws}}, \code{\link[priorsense:powerscale-sensitivity]{powerscale_sensitivity}}}

  \item{stats}{\code{\link[stats]{coef}}, \code{\link[stats:fitted.values]{fitted}}, \code{\link[stats]{logLik}}, \code{\link[stats]{predict}}, \code{\link[stats]{terms}}}

  \item{universals}{\code{\link[universals]{estimates}}}
}}

