% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/coef.R, R/estimates.R, R/glance.R,
%   R/logLik.R, R/predict.R, R/terms.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{coef}
\alias{estimates}
\alias{glance}
\alias{logLik}
\alias{fitted}
\alias{predict}
\alias{terms}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{stats}{\code{\link[stats]{coef}}, \code{\link[stats:fitted.values]{fitted}}, \code{\link[stats]{logLik}}, \code{\link[stats]{predict}}, \code{\link[stats]{terms}}}

  \item{universals}{\code{\link[universals]{estimates}}}
}}

