% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.lmb_analysis}
\alias{residuals.lmb_analysis}
\title{Residuals}
\usage{
\method{residuals}{lmb_analysis}(object, ...)
}
\arguments{
\item{object}{The MB analysis object.}

\item{...}{Unused.}
}
\value{
The analysis data set with the standardised residual values.
}
\description{
Extract standardised (pearson) residual values for an LMB analysis.
}
