% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.mb_analysis}
\alias{residuals.mb_analysis}
\title{Residuals}
\usage{
\method{residuals}{mb_analysis}(object, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{...}{Unused.}
}
\value{
The analysis data set with the fitted values.
}
\description{
Extract residual values for an MB analysis.
}

