% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{rsquared}
\alias{rsquared}
\title{R-Squared}
\usage{
rsquared(analysis, observed, ...)
}
\arguments{
\item{analysis}{The analysis object.}

\item{observed}{A string specifying the observed column in the data.}

\item{...}{Unused.}
}
\description{
Calculate the proportion of the variance in the observed data explained
by the fit.
}

