% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Summarize model sensitivity}
\usage{
sensitivity(x, by = "term", ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{by}{A string indicating whether to determine by "term", "parameter", or
"all".}

\item{...}{Additional arguments passed to
\code{\link[priorsense:powerscale-sensitivity]{priorsense::powerscale_sensitivity()}}.}
}
\value{
A dataframe summarizing the sensitivity of the analysis object.
}
\description{
Summarize model sensitivity
}
