% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-analysis-mode.R
\name{set_analysis_mode}
\alias{set_analysis_mode}
\title{Set Analysis Mode}
\usage{
set_analysis_mode(mode = "report")
}
\arguments{
\item{mode}{A string of the analysis mode.}
}
\description{
Sets analysis mode. Possible values are 'debug', 'quick', 'report' or 'paper'.
}
