% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-residuals.R
\name{simulate_residuals}
\alias{simulate_residuals}
\title{Simulate Residuals}
\usage{
simulate_residuals(x)
}
\arguments{
\item{x}{The MB analysis object.}
}
\value{
An mcmc_data of the simulated residuals.
}
\description{
Requires that new_expr includes code like `residual <- res_bern(`
or `residual[i] <- res_norm(` without the `simulate` argument.
}
