% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terms.mb_analysis}
\alias{terms.mb_analysis}
\title{Terms}
\usage{
\method{terms}{mb_analysis}(x, param_type = "fixed",
  include_constant = TRUE, ...)
}
\arguments{
\item{x}{The mb_analysis object.}

\item{param_type}{A flag specifying whether 'fixed', 'random' or 'derived' terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{...}{Not used.}
}
\description{
terms
}
